Implement OpaqueExistentialContainer
Which is a data structure used in the context of passing an argument at
runtime for a parameter whose type is a protocol. This data structure
holds the value along with its type metadata and protocol value witnesses.
The value could be inline, out-of-line (when it's larger than 3 registers)
or a class instance. We still need to handle two possibilities: one where
the protocol is a class-bound protocol, in which case the data structure
being used is slightly different (ClassExistentialContainer). The second
case is where a protocol composition is being used, in which case there
will be multiple value witness tables.