Refactor type registration logic
We now use a cleaner SwiftModule abstraction that leaves the loading and
caching of types up to the registry. This had to be done because it turns
out that one binary can include types that belong to different Swift
modules (which we retrieve using `getModuleContext()`.) This "splitting"
of logical modules seems to be some Apple internal shenanigans.