Implement Swift.Interceptor API
This extends the original Interceptor API in Frida. It basically parses
`args` and passes the Swift version of them to the callbacks. The Swift
version is either a `ObjectInstance`, a `StructValue` or an `EnumValue`
object. It also works with protocol compositions, by projecting whatever
value is in the container, just like the compiler does.