Fix #4
Where `CSSymbolicatorCreateWithPid()` returns a NULL ref for some
reason, so we added `CSSymbolicatorCreateWithTask()` as a fallback. This
commit also adds proper checks to `Swif.available` so it would return
false for cases like this one (e.g. we fail to create the symbolicator
or one of the private APIs is missing.)